Break the huge map generation stuff into three chunks. 1. Tile Layout 2. Map Markers / Spawn Points 3. Collider size and location
