+++ src/map/helpers/instance_generation/mod.rs
@@ -1,0 +1,5 @@
+mod collider_generator;
+mod map_generator;
+mod tile_generator;
+
+pub use map_generator::generate_map_layout;
