+++ src/map/helpers/mod.rs
@@ -1,2 +1,2 @@
-pub mod map_layout;
 pub mod hub_map_layout;
+pub mod instance_generation;
